v1.5.2

+++ cmd/install_pull.go
@@ -14,8 +14,6 @@ import (
 
 type OptionsInstallPull struct {
 	*OptionsCommon `yaml:"optionsCommon" json:"optionsCommon" bson:"optionsCommon" validate:""`
-	Mode           string `yaml:"mode" json:"mode" bson:"mode" validate:""`
-	Runtime        string `yaml:"runtime" json:"runtime" bson:"runtime" validate:""`
 	FileName       string `yaml:"fileName" json:"fileName" bson:"fileName" validate:""`
 	DownloadAgain  bool   `yaml:"downloadAgain" json:"downloadAgain" bson:"downloadAgain" validate:""`
 }
@@ -33,11 +31,14 @@ func NewCmdInstallPull() *cobra.Command {
 	msgUse := fmt.Sprintf("pull")
 	msgShort := fmt.Sprintf("pull and build all container images")
 	msgLong := fmt.Sprintf(`pull and build all container images required for installation`)
-	msgExample := fmt.Sprintf(`  # pull and build all container images required for installing dory in kubernetes over containerd runtime
-  %s install pull --mode kubernetes --runtime containerd -f install-config.yaml
+	msgExample := fmt.Sprintf(`  # if install or use harbor as image repository it will pull and build images
+  # if install nexus it will download nexus init data 
 
-  # pull and build all container images required for installing dory in docker
-  %s install pull --mode docker -f install-config.yaml`, baseName, baseName)
+  # pull and build all container images required for installing dory
+  %s install pull -f install-config.yaml
+
+  # pull and build all container images required for installing dory
+  %s install pull -f install-config.yaml`, baseName, baseName)
 
 	cmd := &cobra.Command{
 		Use:                   msgUse,
@@ -51,10 +52,8 @@ func NewCmdInstallPull() *cobra.Command {
 		},
 	}
 
-	cmd.Flags().StringVar(&o.Mode, "mode", "", "install dory in docker or kubernetes, options: docker, kubernetes")
-	cmd.Flags().StringVar(&o.Runtime, "runtime", "", "dory managed kubernetes cluster's container runtime, options: docker, containerd, crio")
 	cmd.Flags().StringVarP(&o.FileName, "file", "f", "", "install settings YAML file")
-	cmd.Flags().BoolVarP(&o.DownloadAgain, "download-again", "d", false, "download nexus init data and trivy db again even file exists")
+	cmd.Flags().BoolVarP(&o.DownloadAgain, "download-again", "d", false, "download nexus init data again even file exists")
 
 	CheckError(o.Complete(cmd))
 	return cmd
@@ -68,25 +67,6 @@ func (o *OptionsInstallPull) Complete(cmd *cobra.Command) error {
 		return err
 	}
 
-	err = cmd.RegisterFlagCompletionFunc("mode", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		return []string{"kubernetes", "docker"}, cobra.ShellCompDirectiveNoFileComp
-	})
-	if err != nil {
-		return err
-	}
-
-	err = cmd.RegisterFlagCompletionFunc("runtime", func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		return []string{"docker", "containerd", "crio"}, cobra.ShellCompDirectiveNoFileComp
-	})
-	if err != nil {
-		return err
-	}
-
-	err = cmd.MarkFlagRequired("mode")
-	if err != nil {
-		return err
-	}
-
 	return err
 }
 
@@ -96,18 +76,6 @@ func (o *OptionsInstallPull) Validate(args []string) error {
 	err = o.GetOptionsCommon()
 	if err != nil {
 		return err
-	}
-
-	if o.Mode != "docker" && o.Mode != "kubernetes" {
-		err = fmt.Errorf("--mode must be docker or kubernetes")
-		return err
-	}
-
-	if o.Mode == "kubernetes" {
-		if o.Runtime != "docker" && o.Runtime != "containerd" && o.Runtime != "crio" {
-			err = fmt.Errorf("--runtime must be docker, containerd or crio")
-			return err
-		}
 	}
 
 	if o.FileName == "" {
@@ -143,19 +111,8 @@ func (o *OptionsInstallPull) Run(args []string) error {
 		return err
 	}
 
-	var runtime string
-	if o.Mode == "docker" {
-		runtime = "docker"
-	} else if o.Mode == "kubernetes" {
-		runtime = o.Runtime
-	}
-	if runtime == "" {
-		err = fmt.Errorf("--runtime must be docker, containerd or crio")
-		return err
-	}
-
 	var cmdImagePull, cmdImageBuild, cmdImagePullArm64 string
-	switch runtime {
+	switch installConfig.Kubernetes.Runtime {
 	case "docker":
 		cmdImagePull = "docker pull"
 		cmdImageBuild = "docker build"
@@ -171,58 +128,27 @@ func (o *OptionsInstallPull) Run(args []string) error {
 	}
 
 	var dockerImages pkg.InstallDockerImages
-	dockerImages, err = pkg.GetDockerImages(installConfig)
-	if err != nil {
-		return err
-	}
-
 	dockerFileDir := fmt.Sprintf("dory-docker-files")
-	_ = os.RemoveAll(dockerFileDir)
-	_ = os.MkdirAll(dockerFileDir, 0700)
 	dockerFileTplDir := "docker-files"
+	if installConfig.Dory.ImageRepo.Type == "harbor" {
+		dockerImages, err = pkg.GetDockerImages(installConfig)
+		if err != nil {
+			return err
+		}
 
-	for _, dockerImage := range dockerImages.InstallDockerImages {
-		if dockerImage.DockerFile != "" {
-			arr := strings.Split(dockerImage.Source, ":")
-			var tagName string
-			if len(arr) == 2 {
-				tagName = arr[1]
-			} else {
-				tagName = "latest"
-			}
-			dockerFileName := fmt.Sprintf("%s/%s-%s", dockerFileDir, dockerImage.DockerFile, tagName)
+		_ = os.RemoveAll(dockerFileDir)
+		_ = os.MkdirAll(dockerFileDir, 0700)
 
-			bs, err = pkg.FsInstallScripts.ReadFile(fmt.Sprintf("%s/%s/%s", pkg.DirInstallScripts, dockerFileTplDir, dockerImage.DockerFile))
-			if err != nil {
-				err = fmt.Errorf("create %s error: %s", dockerFileName, err.Error())
-				return err
-			}
-			vals := map[string]interface{}{
-				"source":  dockerImage.Source,
-				"tagName": tagName,
-				"target":  dockerImage.Target,
-				"isArm64": false,
-			}
-			strDockerfile, err := pkg.ParseTplFromVals(vals, string(bs))
-			if err != nil {
-				err = fmt.Errorf("create %s error: %s", dockerFileName, err.Error())
-				return err
-			}
-			err = os.WriteFile(fmt.Sprintf("%s", dockerFileName), []byte(strDockerfile), 0600)
-			if err != nil {
-				err = fmt.Errorf("create values.yaml error: %s", err.Error())
-				return err
-			}
-
-			if dockerImage.Arm64 != "" {
-				arr := strings.Split(dockerImage.Arm64, ":")
+		for _, dockerImage := range dockerImages.InstallDockerImages {
+			if dockerImage.DockerFile != "" {
+				arr := strings.Split(dockerImage.Source, ":")
 				var tagName string
 				if len(arr) == 2 {
 					tagName = arr[1]
 				} else {
 					tagName = "latest"
 				}
-				dockerFileName := fmt.Sprintf("%s/%s-%s-arm64v8", dockerFileDir, dockerImage.DockerFile, tagName)
+				dockerFileName := fmt.Sprintf("%s/%s-%s", dockerFileDir, dockerImage.DockerFile, tagName)
 
 				bs, err = pkg.FsInstallScripts.ReadFile(fmt.Sprintf("%s/%s/%s", pkg.DirInstallScripts, dockerFileTplDir, dockerImage.DockerFile))
 				if err != nil {
@@ -230,10 +156,10 @@ func (o *OptionsInstallPull) Run(args []string) error {
 					return err
 				}
 				vals := map[string]interface{}{
-					"source":  dockerImage.Arm64,
+					"source":  dockerImage.Source,
 					"tagName": tagName,
 					"target":  dockerImage.Target,
-					"isArm64": true,
+					"isArm64": false,
 				}
 				strDockerfile, err := pkg.ParseTplFromVals(vals, string(bs))
 				if err != nil {
@@ -245,161 +171,185 @@ func (o *OptionsInstallPull) Run(args []string) error {
 					err = fmt.Errorf("create values.yaml error: %s", err.Error())
 					return err
 				}
+
+				if dockerImage.Arm64 != "" {
+					arr := strings.Split(dockerImage.Arm64, ":")
+					var tagName string
+					if len(arr) == 2 {
+						tagName = arr[1]
+					} else {
+						tagName = "latest"
+					}
+					dockerFileName := fmt.Sprintf("%s/%s-%s-arm64v8", dockerFileDir, dockerImage.DockerFile, tagName)
+
+					bs, err = pkg.FsInstallScripts.ReadFile(fmt.Sprintf("%s/%s/%s", pkg.DirInstallScripts, dockerFileTplDir, dockerImage.DockerFile))
+					if err != nil {
+						err = fmt.Errorf("create %s error: %s", dockerFileName, err.Error())
+						return err
+					}
+					vals := map[string]interface{}{
+						"source":  dockerImage.Arm64,
+						"tagName": tagName,
+						"target":  dockerImage.Target,
+						"isArm64": true,
+					}
+					strDockerfile, err := pkg.ParseTplFromVals(vals, string(bs))
+					if err != nil {
+						err = fmt.Errorf("create %s error: %s", dockerFileName, err.Error())
+						return err
+					}
+					err = os.WriteFile(fmt.Sprintf("%s", dockerFileName), []byte(strDockerfile), 0600)
+					if err != nil {
+						err = fmt.Errorf("create values.yaml error: %s", err.Error())
+						return err
+					}
+				}
 			}
 		}
-	}
-	log.Info(fmt.Sprintf("create docker files in %s success", dockerFileDir))
-	_, _, err = pkg.CommandExec("ls -alh", dockerFileDir)
-	if err != nil {
-		err = fmt.Errorf("create docker files %s error: %s", dockerFileDir, err.Error())
-		return err
-	}
-	time.Sleep(time.Second * 1)
-
-	log.Info("nexus initial data need to download")
-	fmt.Println(fmt.Sprintf("curl -O -L https://doryengine.com/attachments/%s", pkg.NexusInitData))
-
-	log.Info("trivy vulnerabilities database need to download")
-	fmt.Println(fmt.Sprintf("curl -O -L https://doryengine.com/attachments/%s", pkg.TrivyDb))
-
-	log.Info("container images need to pull")
-	for _, idi := range dockerImages.InstallDockerImages {
-		fmt.Println(fmt.Sprintf("%s %s", cmdImagePull, idi.Source))
-		if idi.Arm64 != "" {
-			fmt.Println(fmt.Sprintf("%s %s %s", cmdImagePull, cmdImagePullArm64, idi.Arm64))
+		log.Info(fmt.Sprintf("create docker files in %s success", dockerFileDir))
+		_, _, err = pkg.CommandExec("ls -alh", dockerFileDir)
+		if err != nil {
+			err = fmt.Errorf("create docker files %s error: %s", dockerFileDir, err.Error())
+			return err
 		}
+		time.Sleep(time.Second * 1)
 	}
 
-	log.Info("container images need to build")
-	log.Warning(fmt.Sprintf("all docker files in %s folder, if your machine is without internet connection, build container images by manual", dockerFileDir))
-	for _, idi := range dockerImages.InstallDockerImages {
-		if idi.DockerFile != "" {
-			arr := strings.Split(idi.Source, ":")
-			var tagName string
-			if len(arr) == 2 {
-				tagName = arr[1]
-			} else {
-				tagName = "latest"
-			}
-			fmt.Println(fmt.Sprintf("%s -t %s -f %s/%s-%s %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir))
+	if installConfig.Dory.ArtifactRepo.Type == "nexus" {
+		log.Info("nexus initial data need to download")
+		fmt.Println(fmt.Sprintf("curl -O -L https://doryengine.com/attachments/%s", pkg.NexusInitData))
+	}
+
+	if installConfig.Dory.ImageRepo.Type == "harbor" {
+		log.Info("container images need to pull")
+		for _, idi := range dockerImages.InstallDockerImages {
+			fmt.Println(fmt.Sprintf("%s %s", cmdImagePull, idi.Source))
 			if idi.Arm64 != "" {
-				arr := strings.Split(idi.Arm64, ":")
+				fmt.Println(fmt.Sprintf("%s %s %s", cmdImagePull, cmdImagePullArm64, idi.Arm64))
+			}
+		}
+
+		log.Info("container images need to build")
+		log.Warning(fmt.Sprintf("all docker files in %s folder, if your machine is without internet connection, build container images by manual", dockerFileDir))
+		for _, idi := range dockerImages.InstallDockerImages {
+			if idi.DockerFile != "" {
+				arr := strings.Split(idi.Source, ":")
 				var tagName string
 				if len(arr) == 2 {
 					tagName = arr[1]
 				} else {
 					tagName = "latest"
 				}
-				fmt.Println(fmt.Sprintf("%s -t %s-arm64v8 -f %s/%s-%s-arm64v8 %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir))
+				fmt.Println(fmt.Sprintf("%s -t %s -f %s/%s-%s %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir))
+				if idi.Arm64 != "" {
+					arr := strings.Split(idi.Arm64, ":")
+					var tagName string
+					if len(arr) == 2 {
+						tagName = arr[1]
+					} else {
+						tagName = "latest"
+					}
+					fmt.Println(fmt.Sprintf("%s -t %s-arm64v8 -f %s/%s-%s-arm64v8 %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir))
+				}
 			}
 		}
 	}
 
-	log.Warning("Make sure current host can connect internet, are you sure download neuxs initial data, trivy vulnerabilities database and pull container images now? [YES/NO]")
-	reader := bufio.NewReader(os.Stdin)
-	userInput, _ := reader.ReadString('\n')
-	userInput = strings.Trim(userInput, "\n")
-	if userInput != "YES" {
-		err = fmt.Errorf("user cancelled")
-		return err
-	}
+	if installConfig.Dory.ImageRepo.Type == "harbor" || installConfig.Dory.ArtifactRepo.Type == "nexus" {
+		arr := []string{}
+		if installConfig.Dory.ImageRepo.Type == "harbor" {
+			arr = append(arr, "pull container images")
+		}
+		if installConfig.Dory.ArtifactRepo.Type == "nexus" {
+			arr = append(arr, "download nexus initial data")
+		}
+		log.Warning(fmt.Sprintf("make sure current host can connect internet, are you sure %s now? [YES/NO]", strings.Join(arr, " and ")))
+		reader := bufio.NewReader(os.Stdin)
+		userInput, _ := reader.ReadString('\n')
+		userInput = strings.Trim(userInput, "\n")
+		if userInput != "YES" {
+			err = fmt.Errorf("user cancelled")
+			return err
+		}
 
-	var isDownloadNexus bool
-	fi, err := os.Stat(pkg.NexusInitData)
-	if err == nil {
-		if !fi.IsDir() {
-			if o.DownloadAgain {
+		var isDownloadNexus bool
+		if installConfig.Dory.ArtifactRepo.Type == "nexus" {
+			fi, err := os.Stat(pkg.NexusInitData)
+			if err == nil {
+				if !fi.IsDir() {
+					if o.DownloadAgain {
+						isDownloadNexus = true
+					}
+				} else {
+					_ = os.RemoveAll(pkg.NexusInitData)
+					isDownloadNexus = true
+				}
+			} else {
 				isDownloadNexus = true
 			}
-		} else {
-			_ = os.RemoveAll(pkg.NexusInitData)
-			isDownloadNexus = true
 		}
-	} else {
-		isDownloadNexus = true
-	}
-	baseName := pkg.GetCmdBaseName()
-	if isDownloadNexus {
-		log.Info(fmt.Sprintf("start to download nexus initial data"))
-		_, _, err = pkg.CommandExec(fmt.Sprintf("curl -O -L https://doryengine.com/attachments/%s", pkg.NexusInitData), ".")
-		if err != nil {
-			err = fmt.Errorf("download nexus initial data error: %s", err.Error())
-			return err
-		}
-		log.Info(fmt.Sprintf("download nexus initial data %s success", pkg.NexusInitData))
-		log.Info(fmt.Sprintf("if run '%s install run [options]' or '%s install script [options]' command, make sure run this command at %s file's directory", baseName, baseName, pkg.NexusInitData))
-	}
-
-	var isDownloadTrivy bool
-	fi, err = os.Stat(pkg.TrivyDb)
-	if err == nil {
-		if !fi.IsDir() {
-			if o.DownloadAgain {
-				isDownloadTrivy = true
-			}
-		} else {
-			_ = os.RemoveAll(pkg.TrivyDb)
-			isDownloadTrivy = true
-		}
-	} else {
-		isDownloadTrivy = true
-	}
-	if isDownloadTrivy {
-		log.Info(fmt.Sprintf("start to download trivy vulnerabilities database"))
-		_, _, err = pkg.CommandExec(fmt.Sprintf("curl -O -L https://doryengine.com/attachments/%s", pkg.TrivyDb), ".")
-		if err != nil {
-			err = fmt.Errorf("download trivy vulnerabilities database error: %s", err.Error())
-			return err
-		}
-		log.Info(fmt.Sprintf("download trivy vulnerabilities database %s success", pkg.TrivyDb))
-		log.Info(fmt.Sprintf("if run '%s install run [options]' or '%s install script [options]' command, make sure run this command at %s file's directory", baseName, baseName, pkg.TrivyDb))
-	}
-
-	log.Info("pull and build container images begin")
-	for i, idi := range dockerImages.InstallDockerImages {
-		_, _, err = pkg.CommandExec(fmt.Sprintf("%s %s", cmdImagePull, idi.Source), ".")
-		if err != nil {
-			err = fmt.Errorf("pull container image %s error: %s", idi.Source, err.Error())
-			return err
-		}
-		if idi.Arm64 != "" {
-			_, _, err = pkg.CommandExec(fmt.Sprintf("%s %s %s", cmdImagePull, cmdImagePullArm64, idi.Arm64), ".")
+		baseName := pkg.GetCmdBaseName()
+		if isDownloadNexus {
+			log.Info(fmt.Sprintf("start to download nexus initial data"))
+			_, _, err = pkg.CommandExec(fmt.Sprintf("curl -O -L https://doryengine.com/attachments/%s", pkg.NexusInitData), ".")
 			if err != nil {
-				err = fmt.Errorf("pull container image %s error: %s", idi.Arm64, err.Error())
+				err = fmt.Errorf("download nexus initial data error: %s", err.Error())
 				return err
 			}
+			log.Info(fmt.Sprintf("download nexus initial data %s success", pkg.NexusInitData))
+			log.Info(fmt.Sprintf("if run '%s install run [options]' or '%s install script [options]' command, make sure run this command at %s file's directory", baseName, baseName, pkg.NexusInitData))
 		}
-		if idi.DockerFile != "" {
-			arr := strings.Split(idi.Source, ":")
-			var tagName string
-			if len(arr) == 2 {
-				tagName = arr[1]
-			} else {
-				tagName = "latest"
-			}
-			_, _, err = pkg.CommandExec(fmt.Sprintf("%s -t %s -f %s/%s-%s %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir), ".")
-			if err != nil {
-				err = fmt.Errorf("build container image %s error: %s", idi.Source, err.Error())
-				return err
-			}
-			if idi.Arm64 != "" {
-				arr := strings.Split(idi.Arm64, ":")
-				var tagName string
-				if len(arr) == 2 {
-					tagName = arr[1]
-				} else {
-					tagName = "latest"
-				}
-				_, _, err = pkg.CommandExec(fmt.Sprintf("%s -t %s-arm64v8 -f %s/%s-%s-arm64v8 %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir), ".")
+
+		if installConfig.Dory.ImageRepo.Type == "harbor" {
+			log.Info("pull and build container images begin")
+			for i, idi := range dockerImages.InstallDockerImages {
+				_, _, err = pkg.CommandExec(fmt.Sprintf("%s %s", cmdImagePull, idi.Source), ".")
 				if err != nil {
-					err = fmt.Errorf("build container image %s error: %s", idi.Arm64, err.Error())
+					err = fmt.Errorf("pull container image %s error: %s", idi.Source, err.Error())
 					return err
 				}
+				if idi.Arm64 != "" {
+					_, _, err = pkg.CommandExec(fmt.Sprintf("%s %s %s", cmdImagePull, cmdImagePullArm64, idi.Arm64), ".")
+					if err != nil {
+						err = fmt.Errorf("pull container image %s error: %s", idi.Arm64, err.Error())
+						return err
+					}
+				}
+				if idi.DockerFile != "" {
+					arr := strings.Split(idi.Source, ":")
+					var tagName string
+					if len(arr) == 2 {
+						tagName = arr[1]
+					} else {
+						tagName = "latest"
+					}
+					_, _, err = pkg.CommandExec(fmt.Sprintf("%s -t %s -f %s/%s-%s %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir), ".")
+					if err != nil {
+						err = fmt.Errorf("build container image %s error: %s", idi.Source, err.Error())
+						return err
+					}
+					if idi.Arm64 != "" {
+						arr := strings.Split(idi.Arm64, ":")
+						var tagName string
+						if len(arr) == 2 {
+							tagName = arr[1]
+						} else {
+							tagName = "latest"
+						}
+						_, _, err = pkg.CommandExec(fmt.Sprintf("%s -t %s-arm64v8 -f %s/%s-%s-arm64v8 %s", cmdImageBuild, idi.Target, dockerFileDir, idi.DockerFile, tagName, dockerFileDir), ".")
+						if err != nil {
+							err = fmt.Errorf("build container image %s error: %s", idi.Arm64, err.Error())
+							return err
+						}
+					}
+				}
+				log.Success(fmt.Sprintf("# progress: %d/%d %s", i+1, len(dockerImages.InstallDockerImages), idi.Target))
 			}
+			log.Success(fmt.Sprintf("pull and build container images success"))
 		}
-		log.Success(fmt.Sprintf("# progress: %d/%d %s", i+1, len(dockerImages.InstallDockerImages), idi.Target))
+	} else {
+		log.Success(fmt.Sprintf("make sure current host can connect internet, no container images need to pull now"))
 	}
-	log.Success(fmt.Sprintf("pull and build container images success"))
 
 	defer color.Unset()
 	return err
